chatroom/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
awsebcli==3.15.1
beautifulsoup4==4.7.1
blessed==1.15.0
botocore==1.12.150
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
cffi==1.12.3
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.5
decorator==4.4.0
Django==2.1.7
django-webpush==0.2.7
django-widget-tweaks==1.4.3
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
extraction==0.3
future==0.16.0
hiredis==1.0.0
html5lib==1.0.1
http-ece==1.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
jsonschema==2.6.0
lxml==4.3.3
metadata-parser==0.10.0
msgpack==0.6.1
numpy==1.16.3
pathspec==0.5.9
Pillow==6.0.0
psycopg2==2.8.2
py-vapid==1.5.0
pycparser==2.19
PyHamcrest==1.9.0
python-dateutil==2.8.0
pytz==2018.9
pywebpush==1.6.0
PyYAML==3.13
requests==2.21.0
requests-toolbelt==0.9.1
semantic-version==2.5.0
six==1.12.0
soupsieve==1.9.1
termcolor==1.1.0
texttable==0.9.1
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.3
validators==0.12.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
zope.interface==4.6.0
